+++ public/app/fn-app/index.tsx
@@ -1,0 +1,1 @@
+export * from './fn-dashboard';
