+++ public/app/fn-app/types.ts
@@ -1,0 +1,4 @@
+export interface FNDashboardProps {
+  uid: string;
+  slug: string;
+}
